* Testklassen für die RuleActionQueue und ihre Helfer angelegt, brauch aber noch weitere Regelklassen damit sinnvoll getestet werden kann  * Der ReasonProcessor arbeitet jetzt nach dem immediate delta-Iteration Prinzip  * die Beispiel Ontologie sample.owl kann nach diesem Prinzip verarbeitet werden

+++ src/de/langenmaier/u2r3/util/RuleActionWeightMap.java
@@ -18,9 +18,9 @@ public class RuleActionWeightMap extends HashMap<RuleAction, Double> {
 
 	public Double put(RuleAction key) {
 		if (containsKey(key)) {
-			Double oldWeight = get(key);
+			double oldWeight = get(key);
 			//here is it possible to tweak the strategy
-			key.setWeight(1+oldWeight.doubleValue());
+			key.setWeight(1+oldWeight);
 		}
 		return super.put(key, key.getWeight());
 	}
